Adding prose template, apa csl, and Makefile edits
Adding a prose template that is for simple non-fiction writing, as well as
adding APA style citations to the csl files, which kieran didn't have. Also, I
am changing the tempate make file in the Makefile directory, because it was set
to apsa stlye or apjs style, and not ASA. Now the default is ASA stlye.