Adding UCETD template and chicago-syllabus csl tweaks
I adapted the Uchicago Dissertation latex template into
a pandoc template that can be called on the command Line.
I got this template from https://github.com/mvzink/ucetd-latex and
it depends on a cls file stored on the latex tree at
/Users/timothyelder/texmf/tex/latex/latex-custom-te. I will
need to tweak this template more to work with my bibliography
workflow.

Also tweaked the chicago-syllabus csl as it was making DOIs spill
off the margin. Not sure about a better fix.